pyproject.toml
[project]
name = "newsbot"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "curl-cffi>=0.14.0",
    "markdownify>=1.2.2",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
]
